requirements.txt
pandas
datasets
openpyxl
openai
anthropic
python-dotenv
scikit-learn
hdbscan
plotly
kneed